delete blog and replace with programs outline
