--- prng/blog/another-post/index.mdx
@@ -1,11 +1,0 @@
----
-title: "Another Post"
-date: "2021-07-24"
-slug: "another-post"
-hero_image: "./anthony-duran-eLUBGqKGdE4-unsplash.jpg"
-hero_image_alt: "A grey and white pitbull wading happily in a pool"
-hero_image_credit_text: "Anthony Duran"
-hero_image_credit_link: "https://unsplash.com/photos/eLUBGqKGdE4"
----
-
-Here's another post! It's even better than the first one!
