--- prng/blog/my-first-post/index.mdx
@@ -1,17 +1,0 @@
----
-title: "My First Post"
-date: "2021-07-23"
-slug: "my-first-post"
-hero_image: "./christopher-ayme-ocZ-_Y7-Ptg-unsplash.jpg"
-hero_image_alt: "A gray pitbull relaxing on the sidewalk with its tongue hanging out"
-hero_image_credit_text: "Christopher Ayme"
-hero_image_credit_link: "https://unsplash.com/photos/ocZ-_Y7-Ptg"
----
-
-This is my first blog post! Isn't it *great*?
-
-Some of my **favorite** things are:
-
-* Petting dogs
-* Singing
-* Eating potato-based foods
