--- prng/blog/yet-another-post/index.mdx
@@ -1,11 +1,0 @@
----
-title: "Yet Another Post"
-date: "2021-07-25"
-slug: "yet-another-post"
-hero_image: "./jane-almon-7rriIaBH6JY-unsplash.jpg"
-hero_image_alt: "A white pitbull wearing big googly-eye glasses"
-hero_image_credit_text: "Jane Almon"
-hero_image_credit_link: "https://unsplash.com/photos/7rriIaBH6JY"
----
-
-Wow look at all this content. How do they do it?
