+++ prng/src/pages/about.js
@@ -5,7 +5,7 @@ import Seo from '../components/seo'
 const AboutPage = () => {
   return (
     <Layout pageTitle="About Me">
-      <p>Hi there! I'm the proud creator of this site, which I built with Gatsby.</p>
+      <p>About Page</p>
     </Layout>
   )
 }
