--- prng/src/pages/blog/{mdx.frontmatter__slug}.js
@@ -1,50 +1,0 @@
-import * as React from 'react'
-import { graphql } from 'gatsby'
-import { GatsbyImage, getImage } from 'gatsby-plugin-image'
-import Layout from '../../components/layout'
-import Seo from '../../components/seo'
-
-const BlogPost = ({ data, children }) => {
-  console.log(data);
-  const image = getImage(data.mdx.frontmatter.hero_image)
-
-  return (
-    <Layout pageTitle={data.mdx.frontmatter.title}>
-      <p>{data.mdx.frontmatter.date}</p>
-      <GatsbyImage
-        image={image}
-        alt={data.mdx.frontmatter.hero_image_alt}
-      />
-      <p>
-        Photo Credit:{" "}
-        <a href={data.mdx.frontmatter.hero_image_credit_link}>
-          {data.mdx.frontmatter.hero_image_credit_text}
-        </a>
-      </p>
-      {children}
-    </Layout>
-  )
-}
-
-export const query = graphql`
-  query($id: String) {
-    mdx(id: {eq: $id}) {
-      frontmatter {
-        title
-        date(formatString: "MMMM DD, YYYY")
-        hero_image_alt
-        hero_image_credit_link
-        hero_image_credit_text
-        hero_image {
-          childImageSharp {
-            gatsbyImageData
-          }
-        }
-      }
-    }
-  }
-`
-
-export const Head = ({ data }) => <Seo title={data.mdx.frontmatter.title} />
-
-export default BlogPost
